Don't precalculate `#hash` method
Intentional trade-off. This will make using Value classes
as _hash keys_ a bit slower for repeated hash insertions. But
faster for constructor calls.

Our experience has been that this is definitely the right
trade-off for us: we spend a lot of time in `.new` and `#with`
calls, and very little time inserting Value classes as hash
keys or Set elements.