Create to_hash as an alias of to_h
The JSON serialization code in ActiveSupport::JSON will call `to_hash`
to convert an object before serializing it but will not call `to_h`.
This leads to serializing any internal instance variables of Value
objects, which seems unexpected.

This commit adds a to_hash method which is an alias of to_h so that JSON
serialization will only serialize the underlying hash